call: Hide Inbound/OutboundCall from yarpc package
InboundCall and OutboundCall are APIs for encoding authors. Instead of
exposing them at the top level, this exposes them on /api/encoding.

CallOption and Call, which are service-author-facing APIs still need to be
defined in the same package as Inbound/OutboundCall because they rely on
internal details, so copies of them are exposed at the top-level instead.
