+++ call.go
@@ -22,88 +22,16 @@ package yarpc
 
 import (
 	"context"
-	"errors"
-	"sort"
 
+	"go.uber.org/yarpc/api/encoding"
 	"go.uber.org/yarpc/api/transport"
 )
-
-type keyValuePair struct{ k, v string }
 
 // CallOption defines options that may be passed in at call sites to other
 // services.
 //
 // These may be used to add or alter the request.
-type CallOption struct{ apply func(*OutboundCall) }
-
-// OutboundCall is an outgoing call. It holds per-call options for a request.
-//
-// Encoding authors may use OutboundCall to provide a CallOption-based request
-// customization mechanism, including returning response headers through
-// ResponseHeaders.
-type OutboundCall struct {
-	// request attributes to fill if non-nil
-	headers         []keyValuePair
-	shardKey        *string
-	routingKey      *string
-	routingDelegate *string
-
-	// If non-nil, response headers should be written here.
-	responseHeaders *map[string]string
-}
-
-// NewOutboundCall constructs a new OutboundCall with the given options.
-func NewOutboundCall(options ...CallOption) *OutboundCall {
-	var call OutboundCall
-	for _, opt := range options {
-		opt.apply(&call)
-	}
-	return &call
-}
-
-// WriteToRequest fills the given request with request-specific options from
-// the call.
-//
-// The context MAY be replaced by the OutboundCall.
-func (c *OutboundCall) WriteToRequest(ctx context.Context, req *transport.Request) (context.Context, error) {
-	for _, h := range c.headers {
-		req.Headers = req.Headers.With(h.k, h.v)
-	}
-
-	if c.shardKey != nil {
-		req.ShardKey = *c.shardKey
-	}
-	if c.routingKey != nil {
-		req.RoutingKey = *c.routingKey
-	}
-	if c.routingDelegate != nil {
-		req.RoutingDelegate = *c.routingDelegate
-	}
-
-	// NB(abg): context and error are unused for now but we want to leave room
-	// for CallOptions which can fail or modify the context.
-	return ctx, nil
-}
-
-// ReadFromResponse reads information from the response for this call.
-//
-// This should be called only if the request is unary.
-func (c *OutboundCall) ReadFromResponse(ctx context.Context, res *transport.Response) (context.Context, error) {
-	// We're not using ctx right now but we may in the future.
-	if c.responseHeaders != nil && res.Headers.Len() > 0 {
-		// We make a copy of the response headers because Headers.Items() must
-		// never be mutated.
-		headers := make(map[string]string, res.Headers.Len())
-		for k, v := range res.Headers.Items() {
-			headers[k] = v
-		}
-		*c.responseHeaders = headers
-	}
-
-	// NB(abg): context and error are unused for now but we want to leave room
-	// for CallOptions which can fail or modify the context.
-	return ctx, nil
-}
+type CallOption encoding.CallOption
 
 // ResponseHeaders specifies that headers received in response to this request
 // should replace the given map.
@@ -123,7 +51,7 @@ func (c *OutboundCall) ReadFromResponse(ctx context.Context, res *transport.Resp
 // 	_, ok := headers["hello"]
 // 	fmt.Println(ok)  // false
 func ResponseHeaders(h *map[string]string) CallOption {
-	return CallOption{func(o *OutboundCall) { o.responseHeaders = h }}
+	return CallOption(encoding.ResponseHeaders(h))
 }
 
 // WithHeader adds a new header to the request. Header keys are case
@@ -135,82 +63,22 @@ func ResponseHeaders(h *map[string]string) CallOption {
 // If multiple entries have the same normalized header name, newer entries
 // override older ones.
 func WithHeader(k, v string) CallOption {
-	return CallOption{func(o *OutboundCall) {
-		o.headers = append(o.headers, keyValuePair{k: k, v: v})
-	}}
+	return CallOption(encoding.WithHeader(k, v))
 }
 
 // WithShardKey sets the shard key for the request.
 func WithShardKey(sk string) CallOption {
-	return CallOption{func(o *OutboundCall) { o.shardKey = &sk }}
+	return CallOption(encoding.WithShardKey(sk))
 }
 
 // WithRoutingKey sets the routing key for the request.
 func WithRoutingKey(rk string) CallOption {
-	return CallOption{func(o *OutboundCall) { o.routingKey = &rk }}
+	return CallOption(encoding.WithRoutingKey(rk))
 }
 
 // WithRoutingDelegate sets the routing delegate for the request.
 func WithRoutingDelegate(rd string) CallOption {
-	return CallOption{func(o *OutboundCall) { o.routingDelegate = &rd }}
-}
-
-// InboundCall holds information about the inbound call and its response.
-//
-// Encoding authors may use InboundCall to provide information about the
-// incoming request on the Context. Response headers written to the context
-// using Call.WriteResponseHeader will be recorded and may be written to a
-// ResponseWriter using WriteToResponse.
-type InboundCall struct {
-	resHeaders []keyValuePair
-	req        *transport.Request
-}
-
-type inboundCallKey struct{} // context key for *InboundCall
-
-// NewInboundCall builds a new InboundCall with the given context. A request
-// context is returned and must be used in place of the original.
-//
-// This may be used by encoding authors to provide information about the
-// incoming request on the Context.
-func NewInboundCall(ctx context.Context) (context.Context, *InboundCall) {
-	call := &InboundCall{}
-	return context.WithValue(ctx, inboundCallKey{}, call), call
-}
-
-// getInboundCall returns the inbound call on this context or nil.
-func getInboundCall(ctx context.Context) (*InboundCall, bool) {
-	call, ok := ctx.Value(inboundCallKey{}).(*InboundCall)
-	return call, ok
-}
-
-// ReadFromRequest reads information from the given request.
-//
-// This information may be queried on the context using functions like Caller,
-// Service, Procedure, etc.
-func (ic *InboundCall) ReadFromRequest(req *transport.Request) error {
-	// TODO(abg): Maybe we should copy attributes over so that changes to the
-	// Request don't change the output.
-	ic.req = req
-	return nil
-}
-
-// WriteToResponse writes response information from the InboundCall onto the
-// given ResponseWriter.
-//
-// If used, this must be called before writing the response body to the
-// ResponseWriter.
-func (ic *InboundCall) WriteToResponse(resw transport.ResponseWriter) error {
-	var headers transport.Headers
-	for _, h := range ic.resHeaders {
-		headers = headers.With(h.k, h.v)
-	}
-
-	if headers.Len() > 0 {
-		resw.AddHeaders(headers)
-	}
-
-	return nil
+	return CallOption(encoding.WithRoutingDelegate(rd))
 }
 
 // Call provides information about the current request inside handlers. An
@@ -225,7 +93,7 @@ func (ic *InboundCall) WriteToResponse(resw transport.ResponseWriter) error {
 // 		}
 // 		return response, nil
 // 	}
-type Call struct{ ic *InboundCall }
+type Call encoding.Call
 
 // CallFromContext retrieves information about the current incoming request
 // from the given context. Returns nil if the context is not a valid request
@@ -233,104 +101,57 @@ type Call struct{ ic *InboundCall }
 //
 // The object is valid only as long as the request is ongoing.
 func CallFromContext(ctx context.Context) *Call {
-	if ic, ok := getInboundCall(ctx); ok {
-		return &Call{ic}
-	}
-	return nil
+	return (*Call)(encoding.CallFromContext(ctx))
 }
 
 // WriteResponseHeader writes headers to the response of this call.
 func (c *Call) WriteResponseHeader(k, v string) error {
-	if c == nil {
-		return errors.New(
-			"failed to write response header: " +
-				"Call was nil, make sure CallFromContext was called with a request context")
-	}
-	c.ic.resHeaders = append(c.ic.resHeaders, keyValuePair{k: k, v: v})
-	return nil
+	return (*encoding.Call)(c).WriteResponseHeader(k, v)
 }
 
 // Caller returns the name of the service making this request.
 func (c *Call) Caller() string {
-	if c == nil {
-		return ""
-	}
-	return c.ic.req.Caller
+	return (*encoding.Call)(c).Caller()
 }
 
 // Service returns the name of the service being called.
 func (c *Call) Service() string {
-	if c == nil {
-		return ""
-	}
-	return c.ic.req.Service
+	return (*encoding.Call)(c).Service()
 }
 
 // Procedure returns the name of the procedure being called.
 func (c *Call) Procedure() string {
-	if c == nil {
-		return ""
-	}
-	return c.ic.req.Procedure
+	return (*encoding.Call)(c).Procedure()
 }
 
 // Encoding returns the encoding for this request.
 func (c *Call) Encoding() transport.Encoding {
-	if c == nil {
-		return ""
-	}
-	return c.ic.req.Encoding
+	return (*encoding.Call)(c).Encoding()
 }
 
 // Header returns the value of the given request header provided with the
 // request.
 func (c *Call) Header(k string) string {
-	if c == nil {
-		return ""
-	}
-
-	if v, ok := c.ic.req.Headers.Get(k); ok {
-		return v
-	}
-
-	return ""
+	return (*encoding.Call)(c).Header(k)
 }
 
 // HeaderNames returns a sorted list of the names of user defined headers
 // provided with this request.
 func (c *Call) HeaderNames() []string {
-	if c == nil {
-		return nil
-	}
-
-	var names []string
-	for k := range c.ic.req.Headers.Items() {
-		names = append(names, k)
-	}
-	sort.Strings(names)
-	return names
+	return (*encoding.Call)(c).HeaderNames()
 }
 
 // ShardKey returns the shard key for this request.
 func (c *Call) ShardKey() string {
-	if c == nil {
-		return ""
-	}
-	return c.ic.req.ShardKey
+	return (*encoding.Call)(c).ShardKey()
 }
 
 // RoutingKey returns the routing key for this request.
 func (c *Call) RoutingKey() string {
-	if c == nil {
-		return ""
-	}
-	return c.ic.req.RoutingKey
+	return (*encoding.Call)(c).RoutingKey()
 }
 
 // RoutingDelegate returns the routing delegate for this request.
 func (c *Call) RoutingDelegate() string {
-	if c == nil {
-		return ""
-	}
-	return c.ic.req.RoutingDelegate
+	return (*encoding.Call)(c).RoutingDelegate()
 }
